+++ src-docs/src/views/datagrid/datagrid_example.js
@@ -26,6 +26,7 @@ import {
   EuiDataGridStyle,
   EuiDataGridToolBarVisibilityOptions,
   EuiDataGridColumnVisibility,
+  EuiDataGridColumnActions,
   EuiDataGridPopoverContentProps,
   EuiDataGridControlColumn,
   EuiDataGridToolBarVisibilityColumnSelectorOptions,
@@ -41,9 +42,13 @@ const gridSnippet = `
     // Required. There are 200 total records.
     rowCount={200}
     // Required. Sets up three columns, the last of which has a custom schema we later define down below.
-    // The second column B won't allow clicking in to see the content in a popup.
-    // The first column defines a starting width of 150px and prevents the user from resizing it
-    columns={[{ id: 'A', initialWidth: 150, isResizable: false }, { id: 'B', isExpandable: false }, {id: 'C', schema: 'franchise'}]}
+    // The second column B won't allow clicking in to see the content in a popup and doesn't show move actions in column header cell
+    // The first column defines a starting width of 150px, prevents the user from resizing it and no actions are displayed
+    columns={[
+        { id: 'A', initialWidth: 150, isResizable: false, actions: false },
+        { id: 'B', isExpandable: false, actions: { showMoveLeft: false, showMoveRight: false } },
+        { id: 'C', schema: 'franchise'}
+    ]}
     // Optional. This allows you to initially hide columns. Users can still turn them on.
     columnVisibility={{
       visibleColumns: ['A', 'C'],
@@ -354,6 +359,7 @@ export const DataGridExample = {
         EuiDataGrid,
         EuiDataGridColumn,
         EuiDataGridColumnVisibility,
+        EuiDataGridColumnActions,
         EuiDataGridControlColumn,
         EuiDataGridInMemory,
         EuiDataGridPaginationProps,